+++ packages/namerank-sdk/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @namehash/namerank
 
+## 0.7.0
+
+### Minor Changes
+
+- 8fd3dcb: feat: [sc-25941] Custom API Root Path with subfolder
+
+### Patch Changes
+
+- Updated dependencies [8fd3dcb]
+  - @namehash/nameguard@0.9.0
+
+## 0.6.0
+
+### Minor Changes
+
+- 5fa2d0e: use correct api endpoint
+
+## 0.5.0
+
+### Minor Changes
+
+- a1842e6: feat: [sc-25972] Update NameRank SDK documentation
+
+## 0.4.0
+
+### Minor Changes
+
+- ddc211f: remove console.log
+
 ## 0.3.0
 
 ### Minor Changes